Set default and label format
 - add combobox to startup to set default label
   - reacts to label list
 - add combobox to startup to set label format
   - reacts to current labeling mode
 - refactor dialog components to custom widgets
 - add enums for label formats and modes

Closes: #127
Closes: #128
