+++ labelCloud/definitions/label_formats/__init__.py
@@ -1,0 +1,2 @@
+from .object_detection import ObjectDetectionFormat
+from .semantic_segmentation import SemanticSegmentationFormat
