+++ labelCloud/definitions/label_formats/base.py
@@ -1,0 +1,8 @@
+from enum import Enum
+from typing import List
+
+
+class BaseLabelFormat(str, Enum):
+    @classmethod
+    def list(cls) -> List["BaseLabelFormat"]:
+        return [e.value for e in cls]
