+++ labelCloud/definitions/label_formats/semantic_segmentation.py
@@ -1,0 +1,7 @@
+from typing import List
+
+from .base import BaseLabelFormat
+
+
+class SemanticSegmentationFormat(BaseLabelFormat):
+    BINARY = "binary"
