+++ labelCloud/definitions/types.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from typing import Tuple
 
 from PyQt5.QtGui import QColor
@@ -13,11 +12,6 @@ Translation3D = Point3D
 Dimensions3D = Tuple[float, float, float]  # length, width, height in meters
 
 Color4f = Tuple[float, float, float, float]  # type alias for type hinting
-
-
-class LabelingMode(str, Enum):
-    OBJECT_DETECTION = "object_detection"
-    SEMANTIC_SEGMENTATION = "semantic_segmentation"
 
 
 class Color3f(tuple):
