+++ labelCloud/io/labels/exceptions.py
@@ -12,3 +12,8 @@ class DefaultIdMismatchException(Exception):
 
 class LabelClassNameEmpty(Exception):
     pass
+
+
+class UnknownLabelFormat(Exception):
+    def __init__(self, label_format: str) -> None:
+        super().__init__(f"Unknown label format '{label_format}'.")
