+++ labelCloud/resources/interfaces/settings_interface.ui
@@ -637,7 +637,7 @@ Bounding Box Parameter</string>
       </widget>
      </item>
      <item row="4" column="1">
-      <widget class="QLabel" name="label_standardobjectclass">
+      <widget class="QLabel" name="label_defaultobjectclass">
        <property name="sizePolicy">
         <sizepolicy hsizetype="Preferred" vsizetype="Maximum">
          <horstretch>0</horstretch>
@@ -648,7 +648,7 @@ Bounding Box Parameter</string>
         <string>Default object class for new bounding boxes.</string>
        </property>
        <property name="text">
-        <string>Standard Object Class</string>
+        <string>Default Object Class</string>
        </property>
       </widget>
      </item>
@@ -666,7 +666,7 @@ Bounding Box Parameter</string>
       </widget>
      </item>
      <item row="4" column="2">
-      <widget class="QLineEdit" name="lineEdit_standardobjectclass">
+      <widget class="QComboBox" name="comboBox_defaultobjectclass">
        <property name="toolTip">
         <string>Default object class for new bounding boxes.</string>
        </property>
