+++ labelCloud/view/startup/color_button.py
