+++ setup.cfg
@@ -48,6 +48,7 @@ packages =
     labelCloud.tests
     labelCloud.utils
     labelCloud.view
+    labelCloud.view.startup
 install_requires =
     numpy
     open3d
